requirements.txt
streamlit
streamlit-option-menu
python-dotenv
pandas
langchain-experimental
langchain-community
openai
openpyxl
PyPDF2
python-docx
